+++ dep/acelite/ace/Process.h
@@ -4,7 +4,7 @@
 /**
  *  @file    Process.h
  *
- *  $Id: Process.h 93506 2011-03-09 10:42:51Z vzykov $
+ *  $Id: Process.h 97304 2013-08-29 21:14:43Z shuston $
  *
  *  @author Tim Harrison <harrison@cs.wustl.edu>
  */
@@ -206,7 +206,7 @@ public:
   /// same as the single string in the prior case and can be obtained by
   /// providing max_len. @arg max_len, if non-zero, provides a location
   /// into which the total length of the command line buffer is returned.
-  ACE_TCHAR *command_line_buf (int *max_len = 0);
+  ACE_TCHAR *command_line_buf (size_t *max_len = 0);
 
   /**
    * argv-style command-line options.  Parses and modifies the string
